// akdrilling/src/app/models/user.ts
import { NgModuleDecorator } from '@angular/core';
import { Role } from './role';

export class User {
  role: Role | undefined;
  info: any | undefined;
  ciaSelected: any | undefined;
  name: any | undefined;
  roleDescription: any | undefined;

}


export class Cias {
  info: any | undefined;
}